requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
